Ensure format is in the type that rails expects
Rails expects value of the `:format` option be a symbol. It filters
formats against `Mime::SET.symbols` so formats which are unknown
(or are known but are of an unexpected type) are not used in cache keys
for the Rails' `ActionView::Resolver::Cache`. This made the resolver
cache unusable when there was multiple templates with the same name but
different formats and was the reason the cache was disabled in
rspec-rails [1].

This isn't the only way of fixing this problem- for example Rails could
be changed to ensure symbols aren't compared with strings as was
suggested in [2], but I feel that it's rspec-rails' responsibility
to work with the existing Rails API.

[1] https://github.com/rspec/rspec-rails/pull/864
[2] https://github.com/rails/rails/pull/13461