Speed up view specs by re-enabling resolver cache
The caching was turned off in [1] because different formats were being
cached under the same key. I believe the need to do this was removed
in the previous commit.

This reduces the time it takes to run FutureLearn’s view specs by 50%.

I think it’s sensible to use the view resolver cache during tests because
it’s normally on between requests in other environments including when
creating tests with `ActionView::TestCase` in a new Rails application.
It looks like it was only turned off in rspec-rails to work around this
caching issue.

I haven't added any test coverage in this change since it feels like
it would be testing a negative (the resolver cache isn't cleared).

[1] https://github.com/rspec/rspec-rails/pull/864