Add option to add rails gems to exclusion filters
Rails in backtraces can get quite spammy, and the configuration option
to `filter_gems_from_backtrace` is not well-known since it is not found
in the generator's `rails_helper.rb`.

This adds `filter_rails_from_backtrace!` as a convenience method for
nixing rails gems in the backtrace and adds it to `rails_helper.rb` to
raise awareness.