Continue to improve our nokogiri support specification.
In [this pr](https://github.com/rspec/rspec-rails/pull/1493) we got a
quick fix to our broken builds on travis due to nokogiri 1.6.6.4 and
1.6.6.3 not working on travis. Now that nokogiri is releasing new
versions, we want to continue to get them. This patch ensures we don't
get the versions that don't work on travis, but does give us new
versions.