Cache stub_template resolvers between example groups
The `#stub_template` view spec helper works by prepending an
ActionView::FixtureResolver[1] to the view instance's view_paths. This
intercepts queries for specific template files, providing a canned
response that's used to instantiate a dummy ActionView::Template.

These templates still require compilation and finalization[2], however.
As a new resolver is created for each stub in every example, this means
that a single `stub_template` call in a `before` block that covers 20
examples will result in the same template being compiled and finalized
20 times.

Both compilation and finalization can be expensive, so to avoid this
where possible, this change adds a global resolver cache for stubbed
templates.

[1] https://github.com/rails/rails/blob/9f65d2a08bc80a94bbb2c0b6e00957c7059aed25/actionview/lib/action_view/testing/resolvers.rb#L6-L9
[2] https://github.com/rails/rails/blob/9f65d2a08bc80a94bbb2c0b6e00957c7059aed25/actionview/lib/action_view/template.rb#L118-L126